rename internal/base to internal/-base
